@@ -19,7 +19,7 @@
         "preferred-install": "dist"
     },
     "require": {
-        "php": "^7.3 || ^8.0",
+        "php": "^8.0",
         "illuminate/database": "^8.13 || ^9.0",
         "illuminate/support": "^8.0 || ^9.0",
         "laravel-interaction/support": "^1.0 || ^2.0"
